# asset-service/docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Asset Management Service API
  description: |
    Asset Management Service API
    
    Comprehensive asset lifecycle management including:
    - Asset CRUD operations
    - Master data management (Categories, Makes, Models, Vendors, etc.)
    - User-Asset linking
    - Compliance and validation
    - Document management
    - Warranty and AMC management
    - Audit logging
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8082
    description: Local development server
  - url: https://api.example.com/asset-service
    description: Production server
...
paths:
...
  /api/asset/v1/components:
    get:
      tags:
        - Components
      summary: List all components
      responses:
        '200':
          description: List of components

    post:
      tags:
        - Components
      summary: Create component
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComponentRequest'
      responses:
        '200':
          description: Component created
  /api/asset/v1/components/{id}:
    get:
      tags:
        - Components
      summary: Get component by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Component found

    put:
      tags:
        - Components
      summary: Update component
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComponentRequest'
      responses:
        '200':
          description: Component updated

    delete:
      tags:
        - Components
      summary: Delete component
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Component deleted

  # ============================================================
  # USER LINK ENDPOINTS
  # ============================================================
...
  /api/asset/v1/documents:
    post:
      tags:
        - Documents
      summary: Upload document
      description: Upload a document for an asset or entity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - entityType
                - entityId
              properties:
                file:
                  type: string
                  format: binary
                entityType:
                  type: string
                entityId:
                  type: integer
                  format: int64
                documentType:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
  /api/asset/v1/documents/{id}:
    get:
      tags:
        - Documents
      summary: Get document by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Document found
        '404':
          description: Document not found

    delete:
      tags:
        - Documents
      summary: Delete document
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Document deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResponseWrapper:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
...
    ComponentRequest:
      type: object
      required:
        - componentName
      properties:
        componentName:
          type: string
        description:
          type: string
...